requirements.txt
# Core data engineering
pyspark>=3.5.0
duckdb>=0.9.0
requests>=2.31.0
python-dateutil>=2.8.2

# Dashboarding
streamlit>=1.30.0
pandas>=2.1.0

# dbt
dbt-duckdb>=1.7.0
